backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic[email]==2.10.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
